Crate rustyGFX

Copyright <2018> [This program is licensed under the "Apache Licence"] Please see the file LICENSE in the source distribution of this software for license terms. This program is designed to simply render simple 3D obj objects to the screen and allow the user cycle through objects, add and remove them on the fly, and adjust the vertex and fragment shaders on the fly this main file contains init to initalize the window, run to run the program and draw to render the objects on the screen

Modules

camera
this function was taken from the glium tutorials that can be found at github.com/glium/glium/examples it is unmodified from its original state this file contains the code that controls the camera movement and allows for users to move around in 3D space

mat
Copyright <2018> the following definitions are influenced by the glium examples since they are the format needed for glium they are not an exact match to the structs used in the glium examples but are a re-partioning of the structs used so all credit goes to them. [This program is licensed under the "Apache Licence"] Please see the file LICENSE in the source distribution of this software for license terms. this file contains the structs used in the program to make the math in the program make more sense.

object_parser
Copyright <2018> [This program is licensed under the "Apache Licence"] Please see the file LICENSE in the source distribution of this software for license terms. The object_parser file contains all of the functions used to parse obj files into the matrix_math_helper structs to be used in the main program

shader_parser
Copyright <2018> [This program is licensed under the "Apache Licence"] Please see the file LICENSE in the source distribution of this software for license terms. The shader_parser file contains the code to read in a text file and convert it into a string to be used as a shader

window
Copyright <2018> [This program is licensed under the "Apache Licence"] Please see the file LICENSE in the source distribution of this software for license terms. the window file creates and initializes the window that will be drawn in by the program