Module rustyGFX::camera

this function was taken from the glium tutorials that can be found at github.com/glium/glium/examples it is unmodified from its original state this file contains the code that controls the camera movement and allows for users to move around in 3D space

Functions

view_matrix
[This program is licensed under the "Apache Licence"] Please see the file LICENSE in the source distribution of this software for license terms. function name: view_matrix inputs: position, direction, up outputs: [[f32;4];4] description: this function takes in a position, direction and up of a camera and returns a view matrix