Module rustyGFX::object_parser::mat

Copyright <2018> the following definitions are influenced by the glium examples since they are the format needed for glium they are not an exact match to the structs used in the glium examples but are a re-partioning of the structs used so all credit goes to them. [This program is licensed under the "Apache Licence"] Please see the file LICENSE in the source distribution of this software for license terms. this file contains the structs used in the program to make the math in the program make more sense.

Structs

Face
struct name: Face description: a structure to hold the face data that is recieved from parsing obj files

Normal
struct name: Normal description: a structure to hold the normal data for a face of a 3D object

PosTex
struct name: PosTex description: a structure to hold the positon and texture coordinates of a Vertex

Texture
struct name: Texture description: a structure to hold the texture coordinates for a 3D object

Vertex
struct name: Vertex description: a structure to hold the position of a point in 3D space